create 1 single example
