--- example/group/go.mod
@@ -1,23 +1,0 @@
-module github.com/topicuskeyhub/go-keyhub/example-group
-
-go 1.17
-
-require github.com/topicuskeyhub/go-keyhub v0.2.1
-
-require github.com/google/uuid v1.3.0
-
-require (
-	github.com/coreos/go-oidc v2.2.1+incompatible // indirect
-	github.com/dghubble/sling v1.4.0 // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
-	github.com/google/go-querystring v1.1.0 // indirect
-	github.com/pquerna/cachecontrol v0.1.0 // indirect
-	golang.org/x/crypto v0.0.0-20211117183948-ae814b36b871 // indirect
-	golang.org/x/net v0.0.0-20211118161319-6a13c67c3ce4 // indirect
-	golang.org/x/oauth2 v0.0.0-20211104180415-d3ed0bb246c8 // indirect
-	google.golang.org/appengine v1.6.7 // indirect
-	google.golang.org/protobuf v1.27.1 // indirect
-	gopkg.in/square/go-jose.v2 v2.6.0 // indirect
-)
-
-replace github.com/topicuskeyhub/go-keyhub => ../..
